+++ src/App.tsx
@@ -5,6 +5,7 @@ import NavbarContainer from './components/header/navbar/NavbarContainer';
 import Header from './components/header/Header';
 import useAppContext from './hooks/contexts/useAppContext';
 import SearchBarContainer from './components/search-bar/SearchBarContainer';
+import useCustomEffect from './hooks/useCustomEffect';
 
 const App = () => {
   const { isSmallDevice } = useAppContext();
@@ -16,6 +17,32 @@ const App = () => {
       !location.pathname.includes(NAVIGATION_PATHS.viewer)
     );
   };
+
+  // Scripts for analysis
+  useCustomEffect(() => {
+    // Development
+    if (import.meta.env.ELVIRA_DEV === 'true') {
+      const script = document.createElement('script');
+      script.async = true;
+      script.src = 'https://analytics.elvira.fiit.stuba.sk/script.js';
+      script.setAttribute(
+        'data-website-id',
+        'afe12559-da42-4f71-99f0-18fffc15ba67'
+      );
+      document.head.appendChild(script);
+    }
+    // Production
+    else if (import.meta.env.ELVIRA_PROD === 'true') {
+      const script = document.createElement('script');
+      script.async = true;
+      script.src = 'https://analytics.elvira.fiit.stuba.sk/script.js';
+      script.setAttribute(
+        'data-website-id',
+        '2fb837fb-1927-404a-8154-32df4945927c'
+      );
+      document.head.appendChild(script);
+    }
+  }, []);
 
   return (
     <div className='min-h-screen h-screen flex w-screen bg-white dark:bg-gray text-black dark:text-white overflow-auto'>